Changes made to the infomenu activity layout
Having some troubles with Android development and all
it's possibilities, even though it's an amazing system, i'ts
also quite hard for me and everything I will add.

So here I changed a bit the infomenu layout and this has
brought me a bug, that I hope to have fixed soon.